Attempt to fix the crashes from 2^32 Params
I think that the 2^32 Params in question each defined a different
integrality datum, amazingly enough. The various integrality data
appearing were indexed by an unsigned int int_sys_nr; so those all had
to be changed to unsigned long int. I'll test now.

+++ sources/gkmod/repr.cpp
@@ -330,7 +330,7 @@ RatWeight Rep_context::offset
   RatWeight result = gamlam - srm1.gamma_lambda();
   auto& ic = inner_class();
   InvolutionNbr inv = kgb().inv_nr(srm0.x());
-  unsigned int int_sys_nr;
+  unsigned long int int_sys_nr;
   const auto codec = ic.integrality_codec(gamlam,inv,int_sys_nr);
   result -= theta_1_preimage(result,codec);
   assert((codec.coroots_matrix*result).is_zero());
